fix: resolve int64_t/qsizetype -> QJsonValue ambiguity on LP64 (GCC/Linux)
The new GCC/Linux leg caught a real portability bug: on LP64 Linux int64_t and
qsizetype are `long`, for which QJsonValue's int/qint64/double constructors are
all viable, so the conversion is ambiguous. It only compiled on macOS because
there int64_t is `long long` (== qint64, an exact match). Cast to qint64 at the
QJSON boundary in ReplayManager (telemetry timestamps) and UIManager (source
count). ControlState's *Ms fields are already qint64, so they are unaffected.

+++ recorder_engine/replaymanager.cpp
@@ -73,8 +73,10 @@ bool ReplayManager::recordTelemetryEvent(const QString &feedId, const QJsonObjec
 
         recorded = payload;
         recorded.insert(QStringLiteral("feedId"), feedId);
-        recorded.insert(QStringLiteral("olrReceiveMs"), receiveMs);
-        recorded.insert(QStringLiteral("olrEffectiveMs"), effectiveMs);
+        // Cast to qint64: int64_t is `long` on LP64 Linux, which makes the
+        // QJsonValue conversion ambiguous; qint64 (long long) is the exact match.
+        recorded.insert(QStringLiteral("olrReceiveMs"), static_cast<qint64>(receiveMs));
+        recorded.insert(QStringLiteral("olrEffectiveMs"), static_cast<qint64>(effectiveMs));
         recorded.insert(QStringLiteral("olrTelemetryDelayMs"), delayMs);
 
         m_muxer->writeTelemetryPacket(
